skills/test/ai-test-system/requirements.txt
openpyxl
Pillow
selenium
PyYAML
requests
